I've commented out the old classdiff behaviour, and started using the new classdiff API, which generates a tree of changes, and then walks that tree to generate the output. I will be ripping the commented sections out in the near future. I know it's messy.
I've modified the get_identifier for methods a little bit in an attempt to
work around the ugliness of bridge methods. I've changed it such that the
identifier for a method is once again name(args) for normal methods, but is
name(args):type for bridge methods. This means that changes of return type
will show up correctly (instead of as a remove and an addition), and
hopefully bridge changes will still be visible as well.